Extracted KDNode as separate class
